requirements.txt
appnope==0.1.0
decorator==4.0.10
Django==1.10
django-cors-headers==1.1.0
django-discover-runner==1.0
django-filter==0.14.0
django-role-permissions==1.2
django-widget-tweaks==1.4.1
djangorestframework==3.4.4
djangorestframework-jwt==1.8.0
entrypoints==0.2.2
httpie==0.9.6
ipykernel==4.3.1
ipython==5.0.0
ipython-genutils==0.1.0
ipywidgets==5.2.2
Jinja2==2.8
jsonschema==2.5.1
jupyter==1.0.0
jupyter-client==4.3.0
jupyter-console==5.0.0
jupyter-core==4.1.0
Markdown==2.6.6
MarkupSafe==0.23
mistune==0.7.3
nbconvert==4.2.0
nbformat==4.0.1
notebook==4.2.1
pexpect==4.2.0
pickleshare==0.7.3
prompt-toolkit==1.0.3
psycopg2==2.6.2
ptyprocess==0.5.1
Pygments==2.1.3
PyJWT==1.4.2
pyzmq==15.3.0
qtconsole==4.2.1
requests==2.11.0
simplegeneric==0.8.1
six==1.10.0
terminado==0.6
tornado==4.4.1
traitlets==4.2.2
wcwidth==0.1.7
widgetsnbextension==1.2.6
